Avoid rejection counters overfill

+++ include/microcanonical_sampler.h
@@ -110,8 +110,8 @@ private:
   std::map<std::array<int, 3>, std::vector<double>> thresholds3_, thresholds2_;
   int debug_printout_;
   bool quantum_statistics_;
-  int accepted23_count_ = 0, accepted32_count_ = 0, rejected23_count_ = 0,
-      rejected32_count_ = 0;
+  long long int accepted23_count_ = 0, accepted32_count_ = 0,
+      rejected23_count_ = 0, rejected32_count_ = 0;
 };
 
 /// Convenient printout of the SamplerParticleList
